fix gint_inout
