+++ source/module_hamilt_lcao/module_gint/gint.h
@@ -10,7 +10,7 @@
 //  specific operations for gamma point/multi-k calculations
 
 #include "tools/gint_tools.h"
-#include "tools/gint_inout.hpp"
+#include "tools/gint_inout.h"
 #include "module_cell/module_neighbor/sltk_grid_driver.h"
 #include "module_hamilt_lcao/module_gint/grid_technique.h"
 #include "module_hamilt_lcao/module_hcontainer/hcontainer.h"
